+++ openeogeotrellis/backend.py
@@ -39,7 +39,14 @@ from openeo.internal.process_graph_visitor import ProcessGraphVisitor
 from openeo.metadata import Band, BandDimension, Dimension, SpatialDimension, TemporalDimension
 from openeo.util import TimingLogger, deep_get, dict_no_none, repr_truncate, rfc3339, str_truncate
 from openeo_driver import backend
-from openeo_driver.backend import BatchJobMetadata, ErrorSummary, LoadParameters, OidcProvider, ServiceMetadata
+from openeo_driver.backend import (
+    BatchJobMetadata,
+    ErrorSummary,
+    LoadParameters,
+    OidcProvider,
+    ServiceMetadata,
+    JobListing,
+)
 from openeo_driver.config.load import ConfigGetter
 from openeo_driver.datacube import DriverDataCube, DriverVectorCube
 from openeo_driver.datastructs import SarBackscatterArgs
@@ -452,7 +459,7 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
                             "description": "Specifies the filename prefix when outputting multiple files. By default, depending on the context, 'OpenEO' or a part of the input filename will be used as prefix.",
                             "default": None,
                         },
-                        "separate_asset_per_band": {
+                        "separate_asset_per_band": {  # TODO: should be a boolean that defaults to false
                             "type": "string",
                             "description": "Set to true to write one output tiff per band. If there is a time dimension, the files will be split on time as well.",
                             "default": None,
@@ -1076,7 +1083,17 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
             if isinstance(error, FileNotFoundError):
                 summary = repr_truncate(str(error), width=width)
             else:
-                summary = repr_truncate(error, width=width)
+                if ConfigParams().is_kube_deploy:
+                    # Conditional import just to be sure.
+                    import kubernetes.client.exceptions
+
+                    if isinstance(error, kubernetes.client.exceptions.ApiException):
+                        s = "kubernetes.client.exceptions.ApiException: " + str(error).strip()
+                        summary = repr_truncate(s, width=width)
+                    else:
+                        summary = repr_truncate(error, width=width)
+                else:
+                    summary = repr_truncate(error, width=width)
 
         return ErrorSummary(error, is_client_error, summary)
 
@@ -1577,9 +1594,20 @@ class GpsBatchJobs(backend.BatchJobs):
         else:  # still some running: continue polling
             pass
 
-    def get_user_jobs(self, user_id: str) -> List[BatchJobMetadata]:
+    def get_user_jobs(
+        self,
+        user_id: str,
+        limit: Optional[int] = None,
+        request_parameters: Optional[dict] = None,
+    ) -> Union[List[BatchJobMetadata], JobListing]:
         with self._double_job_registry as registry:
-            return registry.get_user_jobs(user_id=user_id)
+            return registry.get_user_jobs(
+                user_id=user_id,
+                # Make sure to include title (in addition to the basic fields)
+                fields=["title"],
+                limit=limit,
+                request_parameters=request_parameters,
+            )
 
     # TODO: issue #232 we should get this from S3 but should there still be an output dir then?
     def get_job_output_dir(self, job_id: str) -> Path:
@@ -1633,10 +1661,10 @@ class GpsBatchJobs(backend.BatchJobs):
             terrascope_access_token = self._get_terrascope_access_token(user, sentinel_hub_client_alias)
             return self._vault.login_jwt(terrascope_access_token)
 
-        self._start_job(job_id, user, _get_vault_token)
+        self._start_job(job_id, user, _get_vault_token,proxy_user=proxy_user)
 
     def _start_job(self, job_id: str, user: User, get_vault_token: Callable[[str], str],
-                   dependencies: Union[list, None] = None):
+                   dependencies: Union[list, None] = None,proxy_user=None):
         from openeogeotrellis import async_task  # TODO: avoid local import because of circular dependency
 
         user_id = user.user_id
@@ -1990,7 +2018,8 @@ class GpsBatchJobs(backend.BatchJobs):
                 profile=profile,
                 batch_scheduler=get_backend_config().batch_scheduler,
                 yunikorn_queue=get_backend_config().yunikorn_queue,
-                yunikorn_scheduling_timeout=get_backend_config().yunikorn_scheduling_timeout.rstrip()
+                yunikorn_scheduling_timeout=get_backend_config().yunikorn_scheduling_timeout.rstrip(),
+                try_swift_streaming=os.environ.get("TRY_SWIFT_STREAMING"),
             )
 
             if get_backend_config().fuse_mount_batchjob_s3_bucket:
@@ -2088,7 +2117,8 @@ class GpsBatchJobs(backend.BatchJobs):
                     args.append("no_principal")
                     args.append("no_keytab")
 
-                args.append(job_info.get('proxy_user') or user_id)
+                # due to eventual consistency, job_info does not necessarily have the latest info
+                args.append(proxy_user or job_info.get('proxy_user') or user_id)
 
                 if api_version:
                     args.append(api_version)